HSEARCH-5346 Reuse ORM's ClassDetailsRegistry and ignore the BootstrapContext#getJandexView
